requirements.txt
aiogram==3.3.0
aiosqlite==0.19.0
asyncio==3.4.3
fastapi==0.115.8
pydantic==2.5.3
uvicorn==0.34.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1

